seperated __init__ and routes code.
